+++ src/Stack/Dot.hs
@@ -115,7 +115,7 @@ createDependencyGraph :: (StackM env m, HasEnvConfig env, MonadBaseUnlift IO m)
                        => DotOpts
                        -> m (Map PackageName (Set PackageName, DotPayload))
 createDependencyGraph dotOpts = do
-  (_,_,locals,_,sourceMap) <- loadSourceMap NeedTargets defaultBuildOptsCLI
+  (locals,sourceMap) <- loadSourceMap NeedTargets defaultBuildOptsCLI
       { boptsCLITargets = dotTargets dotOpts
       , boptsCLIFlags = dotFlags dotOpts
       }